Fixed Value for HC

+++ Backend/Features/Market/Data/MaterialQuantity.cs
@@ -1,5 +1,4 @@
 ﻿using Mod.DynamicEncounters.Features.Market.Interfaces;
-using Mod.DynamicEncounters.Features.Market.Services;
 
 namespace Mod.DynamicEncounters.Features.Market.Data;
 
@@ -7,7 +6,7 @@ public class MaterialQuantity(long value) : IItemQuantity
 {
     public long Value { get; } = value;
     
-    public long GetReadableValue()
+    public double GetReadableValue()
     {
         return Value >> 24;
     }
